backend/requirements.txt
fastapi
gunicorn
uvicorn
databases[postgresql]
SQLAlchemy==2.0.28
python-dotenv==1.0.1
pydantic[email]
pydantic-settings
httpx>=0.25.0,<0.26.0
h2>=4.0.0,<4.2.0
hyperframe>=6.0.0,<6.1.0
hpack>=4.0.0,<4.1.0
boto3==1.34.42
botocore==1.34.42
aiofiles
beautifulsoup4
lxml
passlib
bcrypt==4.0.1
python-jose[pycryptodome]
psycopg2-binary
asyncpg==0.29.0
python-multipart
boto3
pytz
python-jose[cryptography]
google-auth
requests
redis>=4.5.0
aioredis>=2.0.1